+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+0.21.10:
+  - better validator state when balance not supplied
+
+0.21.9:
+  - enable custom timeouts for POSTs
+
+0.21.8:
+  - remove Lodestar proposals workaround
+  - add client headers for events stream
+
+0.21.7:
+  - use POST for specific validator and validator balance information
+
+0.21.6:
+  - use SSZ on a per-call basis
+
+0.21.5:
+  - ensure POST bodies are logged as JSON
+
 0.21.4:
   - additional nil checks
   - allow non-mainnet configurations